requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic[email]
pwdlib[argon2]
python-jose
python-multipart